dsp,sgf: update CalculateIntermediate3()
It processes 8 data points one by one, and uses the slow _mm_insert_epi8().
Combine two CalculateIntermediate3() together so it could process 16 data points.
Use table lookup to read elements which indices are less than 48.
For elements which indices are larger than 47, since they seldom change values
with the increase of the index, we use comparison and arithmetic operations to
calculate their values.

perf shows the decoder is ~0.6% faster for a 1080p video.

Tried to apply the same updates to CalculateIntermediate5() in cl/334835363.
However, the compiler generates even slower code.

PiperOrigin-RevId: 336152624
Change-Id: Ie4ca24ff9078b676d0498e42d61cca536bc0c75a