neon,convolve: change 2D filter loop orders
Only applies to 2D filter with width 16 and up.
In the horizontal filter step, filter 8 pixels and then jump to the next row
(inner loop is vertical). After filtering the first 8 columns of pixels,
jump to the next 8 pixels in the first row (outer loop is horizontal).
Then during the vertical filter step, its source can be contiguously read.

perf shows the decoder is ~4-7.6% faster.

PiperOrigin-RevId: 342355669
Change-Id: I9b9af3c7d363117a31b45e6a3b4cdc2d8dc47b5c