Add note for reference_frame_index for intra frame
Add a note that the reference_frame_index array is not used when the frame is
an intra frame and therefore may be uninitialized.

The current implementation zero-initializes the entire ObuFrameHeader struct in
the ObuParser class:

  ObuFrameHeader frame_header_ = {};

therefore the reference_frame_index array is all zeros when the frame is an
intra frame. But this should be considered as an implementation detail. Clients
of libgav1 should not depend on this.

PiperOrigin-RevId: 353113421
Change-Id: If62dd273c3004fd67597bb4addee5226b0c75986