Rename the wrapper for Linux gettid() system call
We have been providing a wrapper for the Linux gettid() system call for
for older versions of glibc. The wrapper is named "gettid" so that the
caller can simply call gettid().

A recent change to Chrome (https://crrev.com/c/2654206) changed the
value of __GLIBC_MINOR__ in <features.h> from 31 to 17. That makes
the !__GLIBC_PREREQ(2, 30) test evaluate to true, but <unistd.h> has a
declaration for gettid(), and our gettid() function gets a compilation
error.

Rename our wrapper "GetTid" to avoid this problem in a locally-modified
build environment, where the fake version 2.17 is reported for glibc
2.31 headers.

PiperOrigin-RevId: 354465891
Change-Id: I91f557d953267aa9a864a2784d7459a780ab0b0f