loop_restoration: Use separate stride for border
Do not share strides between buffers that are backed by the
frame buffer and buffers that are used for borders. The former could
be allocated by external frame buffers and could potentially have a
different stride than the latter.

This CL simply changes the DSP functions and keeps the functionality
as-is for ease of review. Follow-up CLs will actually pass the correct
stride into the DSP layer.

PiperOrigin-RevId: 355311318
Change-Id: Ie0962f1d7fec69d988f0a98d37abaa20e2e8ebff