Correct angle range in directional_intrapred_test
Mode_To_Angle[INTRA_MODES] from section 9.3 Conversion Tables on page
406 of the AV1 spec, and
section 6.10.6. Intra frame mode info semantics on page 174, confirm
that the base angles are enumerated with minimum 45 and maximum 203.
The vales of MAX_ANGLE_DELTA and ANGLE_STEP confirm that the angles do
not extend past +/-9 from the base angles. Therefore the minimum angle
is 45-9=36 and the maximum angle is 203+9=212.

PiperOrigin-RevId: 365949013
Change-Id: I5b7c3f51e76333300a76cebc18d82b30d175ed5c