Move use_predicted_segment_id out of BlockParameters
This CL introduces a new structure called BlockCdfContext. Instead
of storing the values for the entire block, this structure will
store the values for the left column and the top rows for each Tile.

For computing the left context, only the current superblock's left
would suffice (maximum of 32 entries). For computing the top context,
the entire row's top context must be stored (32 entries for each
superblock column).

PiperOrigin-RevId: 372001745
Change-Id: I533dedf2e44b05de1f421b27bb34e9b70aa41f60