Move uv_mode into PredictionParameters
Other than the prediction of the current block, uv_mode is used to
infer the is_smooth_prediction while computing intra edge filter
strength for subsequent blocks. So uv_mode is being moved to
PredictionParameters for the former use case. For the latter use
case, a copy of it is made in BlockCdfContext with a special
case for subsampled chroma planes (because of the way it is used
to compute is_smooth_prediction for subsampled chroma planes).

PiperOrigin-RevId: 372148380
Change-Id: I1f9cba38ec8c756dcc4207a4448d9d41e107292b