add kConvolveScaleBorderRight
this is set at 15.

the NEON SIMD code may overread more than 8 bytes into the allocation
padding (or next row) depending on the block size and location. this is
harmless unless the block happens to be at the bottom right, against the
right and bottom border. rather than add additional logic to handle this
case we'll increase the right border size.

this value is used in place of kConvolveBorderRight in
Tile::GetReferenceBlockPosition() to determine whether a block needs to
be extended; fixes MSan issues with reading uninitialized values from
the padding. with a different allocator or row alignment constant this
could have caused a crash.

this clears all remaining MSan errors with NEON builds.

PiperOrigin-RevId: 405454856
Change-Id: I8ab6b64d41c4cce9740a8f7f4a93adb545a03655