trivial: Remove TODO from buffer_pool.cc
BufferPool::GetFreeBuffer function has been marked as thread safe
in the header. It can be called at the same time from both
ObuParser::ParseOneFrame and DecoderImpl::ApplyFilmGrain. To avoid
this race, we will have to move the allocation of film grain frame
buffer into ObuParser which is an overkill. So it is simpler to keep
the current thread safety mechanism in BufferPool::GetFreeBuffer.
PiperOrigin-RevId: 408768458
Change-Id: I0711f45733d03c18166eb6d5a19e05be5d3d3a2f