Remove max_shuffle_height in DirectionalZone2_4xH
The original formula finds the value is always at least 16.

Analysis shows that, for ystep < 151 (75% of Zone2 angles) the
left-column values can always use shuffle when they're going to blend
with top-row values. When ystep >= 151, the max_shuffle_height question
is only relevant for x < 24. I plan to refactor the 8+ width function to
take this into account, and in the meantime this change covers the fact
that 4xH doesn't need to consider it at all.

PiperOrigin-RevId: 446507907
Change-Id: I7213a55829d06de97577cf2608c5eeced6f7c2e2