Simplify the shuffle decision in DirectionalZone2
Each set of 8 columns is sent to an 8xH function that either always
uses shuffle for left column prediction or always uses transpose, in the
segments where the source is neither left-only nor top-only. Analysis finds
that there are only one or two such segments (8x8 sub-blocks) when x > 16.

This eliminates several divide operations, at the cost of using transpose
somewhat more often. However, the excess transposes are constrained to ~25%
of the Zone2 angles, and there only the first 32 columns.

PiperOrigin-RevId: 448063246
Change-Id: I3dfb472a46d1c70a4cf31dde5d9e9709b0bbdc06