PostFilter: further extend right border w/msan
Optimized loop restoration may read into the next row's left border
depending on the start of the last superblock and the size of the right
border. This is safe as the post filter is applied after reconstruction
is complete and the threaded implementations do not read from the left
border.

PiperOrigin-RevId: 480145288
Change-Id: Ia915ebd7a7f09924cb9c4bf550c3fd14f34ea437