obu_parser: Frame header check for new sequence header OBU
According to AV1 spec Section 7.5. Ordering of OBUs, a new coded
video sequence is defined to start at each temporal unit which
satisfies both of the following conditions:
  * A sequence header OBU appears before the first frame header.
  * The first frame header has frame_type equal to KEY_FRAME,
    show_frame equal to 1, show_existing_frame equal to 0, and
    temporal_id equal to 0.

For the consideration of decoding compatibility as libaom does, it
is defined as warning in order to make the decoding job be continued.

Change-Id: Icbc18d6c69c3b9da9ee9ab4ae86e41f60f524055
Reviewed-on: https://chromium-review.googlesource.com/c/codecs/libgav1/+/5369036
Reviewed-by: James Zern <jzern@google.com>
Tested-by: WebM Builds <builds@webmproject.org>
Reviewed-by: Vignesh Venkat <vigneshv@google.com>