threadpool: Do not use gettid on GNU/Hurd.
gettid is not available on GNU/Hurd and not defined in
glibc. It is specific to Linux for its thread
implementation, while GNU/Hurd uses a different one.
Without a definition, compiling libgav1 on GNU/Hurd
will result in an error of undeclared function.

In commit f06328bf, we have already fixed a mistake
usage of a pthread api on GNU/Hurd. The reason why
this error is not noticed might be that abseil has
failed to build on GNU/Hurd for a long time,
therefore libgav1 might have not been compiled once
as well.

Change-Id: Id0a8f8dbd7660d4dbd0f2968b670d439a273fe20
Signed-off-by: Yuqian Yang <crupest@crupest.life>
Reviewed-on: https://chromium-review.googlesource.com/c/codecs/libgav1/+/6239812
Reviewed-by: Vignesh Venkat <vigneshv@google.com>
Reviewed-by: James Zern <jzern@google.com>
Tested-by: WebM Builds <builds@webmproject.org>